+++ Example.Postgre.Filtering/Requests/Models/ExpirationTimeStampShell.cs
@@ -2,6 +2,6 @@ namespace Example.Postgre.Filtering.Requests.Models
 {
     public class ExpirationTimeStampShell
     {
-        public int ExpirationTimeStamp { get; set; }
+        public long ExpirationTimeStamp { get; set; }
     }
 }
