+++ Example.Postgre.Filtering/Requests/Models/FilterShell.cs
@@ -4,7 +4,7 @@ namespace Example.Postgre.Filtering.Requests.Models
     {
         public string FilterHash { get; set; }
         
-        public int ExpirationTimeStamp { get; set; }
+        public long ExpirationTimeStamp { get; set; }
         
         public string FilterExpression { get; set; }
     }
